introduce file-function-service to bootstrap new component to replace menubar-component
