+++ src/app/components/menubar/menubar.component.ts
@@ -1,25 +1,17 @@
-import { combineLatest, Observable } from 'rxjs';
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
-import { first, map, switchMap } from 'rxjs/operators';
+import { first } from 'rxjs/operators';
 import { MatDialog } from '@angular/material/dialog';
 
-import { AlertService } from '../../@core/services/alertsnackbar/altersnackbar.service';
 import { CaBr2Document } from '../../@core/services/loadSave/loadSave.model';
-import { compareArrays } from '../../@core/utils/compare';
 import { DialogFilter } from '../../@core/services/native/tauri.service';
+import FunctionsService from 'src/app/@core/services/file-functions';
 import { GlobalModel } from '../../@core/models/global.model';
 import { IConfigService } from '../../@core/services/config/config.interface';
-import { ILoadSaveService } from '../../@core/services/loadSave/loadSave.interface';
-import { INativeService } from '../../@core/services/native/native.interface';
 import { LocalizedStrings } from '../../@core/services/i18n/i18n.interface';
-import Logger from '../../@core/utils/logger';
 import { ManualComponent } from '../manual/manual.component';
 import { ReportBugComponent } from '../report-bug/report-bug.component';
 import { SettingsComponent } from '../settings/settings.component';
 import TEMPLATES from '../../../assets/docsTemplate.json';
-import { YesNoDialogComponent } from '../yes-no-dialog/yes-no-dialog.component';
-
-const logger = new Logger('menubar');
 
 const DOCS_TEMPLATE = TEMPLATES.docsTemplate;
 
@@ -36,39 +28,17 @@ export class MenubarComponent implements OnInit {
 
   programmVersion!: string;
 
-  private loadFilter: DialogFilter[] = [];
-  private saveFilter: DialogFilter[] = [];
-
   constructor(
     public globals: GlobalModel,
-    private loadSaveService: ILoadSaveService,
-    private nativeService: INativeService,
-    private alertService: AlertService,
     private configService: IConfigService,
     private dialog: MatDialog,
+    private functionService: FunctionsService,
   ) {
     this.globals.localizedStringsObservable.subscribe((strings) => (this.strings = strings));
-
-    this.loadSaveService.getAvailableDocumentTypes().subscribe(
-      (types) => {
-        this.loadFilter = types.load;
-        this.saveFilter = types.save;
-      },
-      (err) => {
-        logger.error('could not get document types:', err);
-        this.alertService.error(this.strings.error.getAvailableDocumentTypes);
-      },
-    );
   }
 
   ngOnInit(): void {
-    this.globals.headerSubject.next({ ...DOCS_TEMPLATE.header });
-    this.globals.humanAndEnvironmentDangerSubject.next(DOCS_TEMPLATE.humanAndEnvironmentDanger);
-    this.globals.rulesOfConductSubject.next(DOCS_TEMPLATE.rulesOfConduct);
-    this.globals.inCaseOfDangerSubject.next(DOCS_TEMPLATE.inCaseOfDanger);
-    this.globals.disposalSubject.next(DOCS_TEMPLATE.disposal);
-
-    this.globals.substanceDataSubject.next([]);
+    this.globals.loadTemplate(DOCS_TEMPLATE);
 
     this.configService
       .getProgramVersion()
@@ -77,7 +47,7 @@ export class MenubarComponent implements OnInit {
   }
 
   newDocument(): void {
-    this.ngOnInit();
+    this.globals.loadTemplate(DOCS_TEMPLATE);
   }
 
   scroll(el: HTMLElement): void {
@@ -88,134 +58,12 @@ export class MenubarComponent implements OnInit {
     this.dialog.open(ReportBugComponent);
   }
 
-  modelToDocument(): Observable<CaBr2Document> {
-    return combineLatest([
-      this.globals.headerObservable,
-      this.globals.substanceDataObservable,
-      this.globals.disposalObservable,
-      this.globals.humanAndEnvironmentDangerObservable,
-      this.globals.inCaseOfDangerObservable,
-      this.globals.rulesOfConductObservable,
-    ]).pipe(
-      map((value) => ({
-        header: value[0],
-        substanceData: value[1],
-        disposal: value[2],
-        humanAndEnvironmentDanger: value[3],
-        inCaseOfDanger: value[4],
-        rulesOfConduct: value[5],
-      })),
-    );
-  }
-
-  documentToModel(doc: CaBr2Document): void {
-    this.globals.headerSubject.next(doc.header);
-    this.globals.substanceDataSubject.next(doc.substanceData);
-    this.globals.disposalSubject.next(doc.disposal);
-    this.globals.humanAndEnvironmentDangerSubject.next(doc.humanAndEnvironmentDanger);
-    this.globals.inCaseOfDangerSubject.next(doc.inCaseOfDanger);
-    this.globals.rulesOfConductSubject.next(doc.rulesOfConduct);
-  }
-
   loadFile(): void {
-    logger.trace('loadFile');
-    this.nativeService
-      .open({
-        filters: this.loadFilter,
-        multiple: false,
-      })
-      .pipe(first())
-      .subscribe(
-        (path) => {
-          this.loadSaveService.loadDocument(path as string).subscribe(
-            (res) => this.documentToModel(res),
-            (err) => {
-              logger.error('loading file failed:', err);
-              this.alertService.error(this.strings.error.loadFile);
-            },
-          );
-        },
-        (err) => logger.trace('open dialog returned error:', err),
-      );
-  }
-
-  /**
-   * Returns `true` if the `CaBr2Document has some unchecked default values
-   */
-  checkUnmodified(document: CaBr2Document): string[] {
-    const unmodified = [];
-    for (const substance of document.substanceData) {
-      if (!substance.checked) {
-        unmodified.push(substance.name.modifiedData ?? substance.name.originalData);
-      }
-    }
-    for (const section of [
-      [
-        document.humanAndEnvironmentDanger,
-        DOCS_TEMPLATE.humanAndEnvironmentDanger,
-        [this.strings.descriptions.humanAndEnvironmentDangerShort],
-      ],
-      [document.rulesOfConduct, DOCS_TEMPLATE.rulesOfConduct, [this.strings.descriptions.rulesOfConductShort]],
-      [document.inCaseOfDanger, DOCS_TEMPLATE.inCaseOfDanger, [this.strings.descriptions.inCaseOfDangerShort]],
-      [document.disposal, DOCS_TEMPLATE.disposal, [this.strings.descriptions.disposalShort]],
-    ]) {
-      if (compareArrays(section[0], section[1])) {
-        unmodified.push(section[2][0]);
-      }
-    }
-    return unmodified;
+    this.functionService.loadFile();
   }
 
   saveFile(type: DialogFilter, document: CaBr2Document): void {
-    // check for development, should never occur in production
-    if (this.saveFilter.includes(type)) {
-      throw Error('unsupported file type');
-    }
-
-    // there should always be exact one fileextension
-    const extension = type.extensions[0];
-
-    this.nativeService
-      .save({ filters: [type] })
-      .pipe(
-        switchMap((filename) => this.loadSaveService.saveDocument(extension, filename as string, document)),
-        first(),
-      )
-      .subscribe(
-        (res) => {
-          logger.debug(res);
-
-          switch (extension) {
-            case 'pdf':
-              this.alertService.success(this.strings.success.exportPDF);
-              break;
-
-            default:
-              this.alertService.success(this.strings.success.saveFile);
-              break;
-          }
-        },
-        (err) => {
-          logger.error(err);
-          // fix for an error that occurs only in windows
-          if (err === 'Could not initialize COM.') {
-            logger.debug('ty windows -.- | attempting fix');
-            this.loadFile();
-            this.saveFile(type, document);
-            return;
-          }
-
-          switch (extension) {
-            case 'pdf':
-              this.alertService.error(this.strings.error.exportPDF);
-              break;
-
-            default:
-              this.alertService.error(this.strings.error.saveFile);
-              break;
-          }
-        },
-      );
+    this.functionService.saveFile(type, document);
   }
 
   openManualDialog(): void {
@@ -237,37 +85,10 @@ export class MenubarComponent implements OnInit {
   }
 
   exportCB2File(): void {
-    this.exportFile({ name: 'CaBr2', extensions: ['cb2'] });
+    this.functionService.exportFile({ name: 'CaBr2', extensions: ['cb2'] }, DOCS_TEMPLATE);
   }
 
   exportPDFFile(): void {
-    this.exportFile({ name: 'PDF', extensions: ['pdf'] });
-  }
-
-  private exportFile(type: DialogFilter): void {
-    logger.trace(`exportFile('${type}')`);
-    this.modelToDocument()
-      .pipe(first())
-      .subscribe((doc) => {
-        const unmodifiedStuff = this.checkUnmodified(doc);
-        if (unmodifiedStuff) {
-          this.dialog
-            .open(YesNoDialogComponent, {
-              data: {
-                iconName: 'warning',
-                title: this.strings.dialogs.unchangedValues.title,
-                content: this.strings.dialogs.unchangedValues.content,
-                listItems: unmodifiedStuff,
-                footerText: this.strings.dialogs.unchangedValues.footer,
-              },
-              panelClass: ['unselectable', 'undragable'],
-            })
-            .afterClosed()
-            .pipe(first())
-            .subscribe((res) => (res ? this.saveFile(type, doc) : undefined));
-        } else {
-          this.saveFile(type, doc);
-        }
-      });
+    this.functionService.exportFile({ name: 'PDF', extensions: ['pdf'] }, DOCS_TEMPLATE);
   }
 }
